#79 readouts
When adding a component to a space, in the panel for
selecting/editing a component, show "readouts," which
are labels and icons that help to distinguish between
component instances that are of the same component
type.

To try this out, add this to the schema.yml for a
component for which you will create a space, then,
after creating the space, edit the component instance
so that its `tag` and `title` properties are defined
and are not empty strings.

schema.yml:

    _targeting:
      -
        property: tag
        icon: tag
      -
        property: title
        icon: clock

then readouts should display when those components
are in the component list for a space.